refactor: move SessionManager to auth/data
