+++ shared/src/commonMain/kotlin/com/example/tasky/common/data/manager/HttpManager.kt
@@ -1,8 +1,8 @@
 package com.example.tasky.common.data.manager
 
 import com.example.tasky.BuildKonfig.API_KEY
+import com.example.tasky.auth.data.manager.SessionManager
 import com.example.tasky.auth.data.model.AccessTokenResponse
-import com.example.tasky.auth.domain.manager.SessionManager
 import com.example.tasky.common.data.model.ErrorResponse
 import com.example.tasky.common.data.util.isSuccess
 import io.ktor.client.HttpClient
